refactor: User와 Article의 데이터베이스 클래스를 패키지에 저장해서 관리
