+++ src/main/java/codesquad/springcafe/user/UserController.java
@@ -1,5 +1,6 @@
 package codesquad.springcafe.user;
 
+import codesquad.springcafe.user.database.UserDatabase;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
