+++ src/main/java/codesquad/springcafe/user/database/UserDatabase.java
@@ -1,5 +1,7 @@
-package codesquad.springcafe.user;
+package codesquad.springcafe.user.database;
 
+import codesquad.springcafe.user.User;
+import codesquad.springcafe.user.UserRequestDto;
 import java.util.List;
 
 public interface UserDatabase {
