+++ src/main/java/codesquad/springcafe/user/database/UserH2Database.java
@@ -1,5 +1,8 @@
-package codesquad.springcafe.user;
+package codesquad.springcafe.user.database;
 
+import codesquad.springcafe.user.User;
+import codesquad.springcafe.user.UserRequestDto;
+import codesquad.springcafe.user.UserRowMapper;
 import java.util.ArrayList;
 import java.util.List;
 import org.springframework.context.annotation.Primary;
