+++ src/test/java/codesquad/springcafe/article/ArticleInMemoryDatabaseTest.java
@@ -2,6 +2,7 @@ package codesquad.springcafe.article;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import codesquad.springcafe.article.database.ArticleInMemoryDatabase;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
