+++ src/test/java/codesquad/springcafe/user/UserInMemoryDatabaseTest.java
@@ -2,6 +2,7 @@ package codesquad.springcafe.user;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import codesquad.springcafe.user.database.UserInMemoryDatabase;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
